+++ packages/protocol/src/index.ts
@@ -47,13 +47,13 @@ export * from 'sprotty-protocol/lib/utils/json';
 export * from 'sprotty-protocol/lib/utils/model-utils';
 // Default export of @eclipse-glsp/protocol
 export * from './action-protocol';
+export * from './client-server-protocol/base-glsp-client';
 export * from './client-server-protocol/glsp-client';
+export * from './client-server-protocol/glsp-server';
 export * from './client-server-protocol/jsonrpc/base-jsonrpc-glsp-client';
 export * from './client-server-protocol/jsonrpc/glsp-jsonrpc-client';
 export * from './client-server-protocol/jsonrpc/glsp-jsonrpc-server';
 export * from './client-server-protocol/jsonrpc/websocket-connection';
-export * from './client-server-protocol/node/glsp-server';
-export * from './client-server-protocol/node/node-glsp-client';
 export * from './client-server-protocol/types';
 export * from './disposable/disposable';
 export * from './model/default-types';
